requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
dataclasses==0.8
google-api-core==1.25.1
google-auth==1.24.0
google-cloud-vision==2.1.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
idna==2.10
libcst==0.3.16
mypy-extensions==0.4.3
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
rsa==4.7
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
